last change

+++ task3/3_Rust.rs
@@ -1,6 +1,6 @@
 use std::io;
 
-fn gcd(mut a: i128, mut b: i128) -> i128 {
+fn gcd(mut a: i64, mut b: i64) -> i64 {
     while b != 0 {
         a %= b;
         std::mem::swap(&mut a, &mut b);
@@ -8,7 +8,7 @@ fn gcd(mut a: i128, mut b: i128) -> i128 {
     a
 }
 
-fn power(base: i128, exp: u32) -> i128 {
+fn power(base: i64, exp: u32) -> i64 {
     if exp == 0 { return 1; }
     base.pow(exp)
 }
@@ -37,27 +37,27 @@ fn main() {
         return;
     }
 
-    let mut c: Vec<i128> = vec![0; (a + 1) as usize];
+    let mut c: Vec<i64> = vec![0; (a + 1) as usize];
     c[1] = 1;
 
     for k in 1..a {
-        let mut next_c = vec![0i128; (a + 1) as usize];
+        let mut next_c = vec![0i64; (a + 1) as usize];
         for i in 1..=k as usize {
-            next_c[i] += (i as i128) * c[i];
+            next_c[i] += (i as i64) * c[i];
             if i + 1 <= a as usize {
-                next_c[i + 1] += (k as i128 + 1 - i as i128) * c[i];
+                next_c[i + 1] += (k as i64 + 1 - i as i64) * c[i];
             }
         }
         c = next_c;
     }
 
-    let mut numerator: i128 = 0;
+    let mut numerator: i64 = 0;
     for i in 1..=a as usize {
-        numerator += c[i] * power(b as i128, (a - i as i32) as u32);
+        numerator += c[i] * power(b as i64, (a - i as i32) as u32);
     }
-    numerator *= b as i128;
+    numerator *= b as i64;
 
-    let denominator = power((b - 1) as i128, (a + 1) as u32);
+    let denominator = power((b - 1) as i64, (a + 1) as u32);
 
     let common = gcd(numerator, denominator);
     let final_num = numerator / common;
