+++ task4/4_c++.cpp
@@ -3,27 +3,27 @@
 
 using namespace std;
 
-int game(const vector<long long> &chain, int n, int m){
+int game(const vector<long> &chain, int n, int m){
     int last_k = 0;
     int index = 0;
 
-    vector<long long> prefix(n + 1, 0);
+    vector<long> prefix(n + 1, 0);
 
     for (int i = 0; i < n; i++) {
         prefix[i + 1] = prefix[i] + chain[i];
     }
-    long long scorePasha = 0, scoreVik = 0;
+    long scorePasha = 0, scoreVik = 0;
     bool pashaTurn = true;
 
     while (index < n) {
         int max_len = min(m, n - index);
         int best_k = -1;
-        long long best_sum = -1e18;
+        long best_sum = -1e18;
         
         for (int k = 1; k <= max_len; k++) {
             if (k == last_k) continue;
 
-            long long sum = prefix[index + k] - prefix[index];
+            long sum = prefix[index + k] - prefix[index];
             if (sum > best_sum || (sum == best_sum && k < best_k)) {
                 best_sum = sum;
                 best_k = k;
@@ -55,7 +55,7 @@ int main(){
     cin >> m;
     cout << "Type in " << n << " numbers" << endl;
 
-    vector<long long> chain_n(n);
+    vector<long> chain_n(n);
 
     for (int i = 0; i < n; i++){
         cin >> chain_n[i];
